requirements.txt
anthropic>=0.40.0
resend>=2.4.0
feedparser>=6.0.11
httpx>=0.27.0
jinja2>=3.1.4
PyYAML>=6.0.2
python-dateutil>=2.9.0
beautifulsoup4>=4.12.3
tzdata>=2024.2
